Fixing rotary encoder

+++ callete-api-impl/src/main/java/callete/api/services/impl/gpio/raspi/RotaryEncoderEventImpl.java
@@ -6,16 +6,16 @@ import callete.api.services.gpio.RotaryEncoderEvent;
  * The event implementation for rotary encoder changes.
  */
 public class RotaryEncoderEventImpl implements RotaryEncoderEvent {
-  private int steps;
+  private long steps;
   private boolean left;
 
-  public RotaryEncoderEventImpl(int steps, boolean left) {
+  public RotaryEncoderEventImpl(long steps, boolean left) {
     this.steps = steps;
     this.left = left;
   }
 
   @Override
-  public int getSteps() {
+  public long getSteps() {
     return steps;
   }
